Create multiple embeddings at once

It is especially useful, for dealing with multiple categorical features.

nn_multi_embedding(
  num_embeddings = "auto",
  embedding_dim = "auto",
  padding_idx = NULL,
  max_norm = NULL,
  norm_type = 2,
  scale_grad_by_freq = FALSE,
  sparse = FALSE,
  .weight = NULL
)

Arguments

num_embeddings
(integer): size of the dictionary of embeddings

embedding_dim
(integer): the size of each embedding vector

padding_idx
(integer, optional): If given, pads the output with the embedding vector at padding_idx (initialized to zeros) whenever it encounters the index.

max_norm
(numeric, optional): If given, each embedding vector with norm larger than max_norm is renormalized to have norm max_norm.

norm_type
(numeric, optional): The p of the p-norm to compute for the max_norm option. Default 2.

scale_grad_by_freq
(logical, optional): If given, this will scale gradients by the inverse of frequency of the words in the mini-batch. Default FALSE

sparse
(logical, optional): If True, gradient w.r.t. weight matrix will be a sparse tensor.

.weight
(torch_tensor or list of torch_tensor) embeddings weights (in case you want to set it manually)

Examples

data("gss_cat", package = "forcats")

gss_cat_transformed <-
  gss_cat %>%
  na.omit() %>%
  sapply(function(x) cat2idx(x)[[1]]) %>%
  as_tibble()

#> Error in cat2idx(x): nie udało się znaleźć funkcji 'cat2idx'

gss_cat_tensor  <- as_tensor(gss_cat_transformed)

#> Error in as_tensor(gss_cat_transformed): nie znaleziono obiektu 'gss_cat_transformed'

.dict_size      <- dict_size(gss_cat_transformed)

#> Error in lapply(X = X, FUN = FUN, ...): nie znaleziono obiektu 'gss_cat_transformed'

.embedding_size <- ceiling(.dict_size ** .25)

#> Error in eval(expr, envir, enclos): nie znaleziono obiektu '.dict_size'

embedding_module <-
  nn_multi_embedding(.dict_size, .embedding_size)

#> Error in nn_multi_embedding(.dict_size, .embedding_size): nie udało się znaleźć funkcji 'nn_multi_embedding'

# Expected output size
sum(.embedding_size)

#> Error in eval(expr, envir, enclos): nie znaleziono obiektu '.embedding_size'

embedding_module(gss_cat_tensor)

#> Error in embedding_module(gss_cat_tensor): nie udało się znaleźć funkcji 'embedding_module'