A prepared subset from M5 Walmart Challenge Dataset in one data frame

A pieced of data cut from the training dataset used in the M5 challenges on Kaggle. M5 is a challenge from a series organized by Spyros Makridakis.

tiny_m5

Format

item_id

The id of the product

dept_id

The id of the department the product belongs to

cat_id

The id of the category the product belongs to

store_id

The id of the store where the product is sold

state_id

The State where the store is located

value

The number of sold units

date

The date in a “y-m-d” format

wm_yr_wk

The id of the week the date belongs to

weekday

The type of the day (Saturday, Sunday, …, Friday)

wday

The id of the weekday, starting from Saturday

month

The month of the date

year

The year of the date

event_name_1

If the date includes an event, the name of this event

event_type_1

If the date includes an event, the type of this event

event_name_2

If the date includes a second event, the name of this event

event_type_2

If the date includes a second event, the type of this event

snap

A binary variable (0 or 1) indicating whether the stores of CA, TX or WI allow SNAP1 purchases on the examined date. 1 indicates that SNAP purchases are allowed

sell_price

he price of the product for the given week/store. The price is provided per week (average across seven days). If not available, this means that the product was not sold during the examined week. Note that although prices are constant at weekly basis, they may change through time (both training and test set)

See also

Examples

# Head of tiny_m5
head(tiny_m5)

#>         item_id   dept_id  cat_id store_id state_id value       date wm_yr_wk
#> 1 HOBBIES_1_158 HOBBIES_1 HOBBIES     CA_1       CA     0 2011-01-29    11101
#> 2 HOBBIES_1_158 HOBBIES_1 HOBBIES     CA_1       CA     0 2011-01-30    11101
#> 3 HOBBIES_1_158 HOBBIES_1 HOBBIES     CA_1       CA     0 2011-01-31    11101
#> 4 HOBBIES_1_158 HOBBIES_1 HOBBIES     CA_1       CA     0 2011-02-01    11101
#> 5 HOBBIES_1_158 HOBBIES_1 HOBBIES     CA_1       CA     0 2011-02-02    11101
#> 6 HOBBIES_1_158 HOBBIES_1 HOBBIES     CA_1       CA     0 2011-02-03    11101
#>     weekday wday month year event_name_1 event_type_1 event_name_2 event_type_2
#> 1  Saturday    1     1 2011                                                    
#> 2    Sunday    2     1 2011                                                    
#> 3    Monday    3     1 2011                                                    
#> 4   Tuesday    4     2 2011                                                    
#> 5 Wednesday    5     2 2011                                                    
#> 6  Thursday    6     2 2011                                                    
#>   snap sell_price
#> 1    0         NA
#> 2    0         NA
#> 3    0         NA
#> 4    1         NA
#> 5    1         NA
#> 6    1         NA