Create a time series dataset object

Create a time series dataset object

ts_dataset(
  data,
  n_timesteps,
  h,
  input_columns = list(x = NULL),
  target_columns = list(y = NULL),
  sample_frac = 1,
  scale = TRUE
)

Arguments

data
(torch_tensor) An input data object

n_timesteps
(integer) Number of timesteps for input tensor

h
(integer) Forecast horizon: number of timesteps for output tensor

input_columns
(list) Output specification

target_columns
(list) Output specification

scale
(logical) Scale feature columns

sample_fram
(numeric) A numeric value > 0. and <= 1 to sample a subset of data

Note

If scale is TRUE, only the input vaiables are scale and not the outcome ones. See: Is it necessary to scale the target value in addition to scaling features for regression analysis? (Cross Validated)

Examples

suppressMessages(library(dplyr))
library(torchts)
data("mts-examples", package="MTS")

ibmspko <-
    ibmspko %>%
    select(date, ibm)

ibm_tensor <- as_tensor(ibmspko, date)

ibm_dataset <-
    ts_dataset(ibm_tensor, n_timesteps = 7, h = 7)

ibm_dataset$.getitem(1)

#> $x
#> torch_tensor
#> (1,.,.) = 
#>  0.01 *
#>   7.2513
#> 
#> (2,.,.) = 
#>  0.01 *
#>   6.2500
#> 
#> (3,.,.) = 
#>  0.01 *
#>   2.9630
#> 
#> (4,.,.) = 
#>  0.01 *
#>   2.7338
#> 
#> (5,.,.) = 
#>  0.01 *
#>   2.7521
#> 
#> (6,.,.) = 
#>  0.01 *
#>  -2.6612
#> 
#> (7,.,.) = 
#>  0.01 *
#>   2.4185
#> [ CPUFloatType{7,1,1} ]
#> 
#> $y
#> torch_tensor
#> (1,.,.) = 
#>  0.01 *
#>   6.7967
#> 
#> (2,.,.) = 
#>  0.01 *
#>   3.5611
#> 
#> (3,.,.) = 
#>  0.01 *
#>   9.2007
#> 
#> (4,.,.) = 
#>  0.01 *
#>  -1.1745
#> 
#> (5,.,.) = 
#>  0.001 *
#>  -1.7240
#> 
#> (6,.,.) = 
#>  0.01 *
#>  -6.3903
#> 
#> (7,.,.) = 
#>  0.001 *
#>  -7.8410
#> [ CPUFloatType{7,1,1} ]
#>